A STM topic model of the inaugural presidential speeches

This package dataset is a stm topic model for the sample_dfm with 25 topics; it was fit to the inaugural presidential speeches in the respective dataset included in the quanteda package (quanteda::data_corpus_inaugural).

sample_topics

Format

a stm object

Details

This dataset is included for convenience and reproducibility; depending on the respective system configuration estimating a topic model can require a long-running process.